Function entropy_get_entropy now in working state
This patch fixes an array indexing error that occurred in the function
entropy_get_entropy. This function is now in a working state. The next
patches will be coming soon, but after a break for development. At this
point in the program, I have not yet completely developed my algorithm.